fix(ui): address agent and chart polish review findings
