+++ package.json
@@ -32,7 +32,7 @@
     "test:public-site-smoke": "bun test scripts/public_site_smoke.test.ts",
     "test:visual-modal": "node scripts/ensure_playwright.mjs && bun scripts/build_test_assets.ts visual-modal && bun test web/components/dashboard/visual-modal-focus.test.ts web/components/dashboard/visual-modal.dom.test.ts",
     "test:catalog-page": "node scripts/ensure_playwright.mjs && bun scripts/build_test_assets.ts catalog-page && bun test web/components/app/catalog-page.dom.test.ts web/components/app/dashboard-icon-picker.test.ts",
-    "test:dashboard-page": "node scripts/ensure_playwright.mjs && bun scripts/build_test_assets.ts dashboard-page && bun test web/components/dashboard/dashboard-page.dom.test.ts && bun test web/components/dashboard/dashboard-page.visuals.dom.test.ts && bun scripts/build_test_assets.ts visualization-host && bun test web/components/dashboard/visualization/host.dom.test.ts",
+    "test:dashboard-page": "node scripts/ensure_playwright.mjs && bun scripts/build_test_assets.ts dashboard-page && bun test web/components/dashboard/dashboard-page.dom.test.ts && bun test web/components/dashboard/dashboard-page.agent.dom.test.ts && bun test web/components/dashboard/dashboard-page.visuals.dom.test.ts && bun scripts/build_test_assets.ts visualization-host && bun test web/components/dashboard/visualization/host.dom.test.ts",
     "test:date-picker": "node scripts/ensure_playwright.mjs && bun scripts/build_test_assets.ts date-picker && bun test web/components/dashboard/filters/date-picker.dom.test.ts",
     "test:dashboard-builder": "node scripts/ensure_playwright.mjs && bun scripts/build_test_assets.ts dashboard-builder && bun test web/components/dashboard/dashboard-builder.dom.test.ts",
     "test:chat-composer": "node scripts/ensure_playwright.mjs && bun scripts/build_test_assets.ts chat-composer && bun test web/components/chat/chat-composer.dom.test.ts",
